+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The builder methods on the Criterion struct now take and return self by value
+  for easier chaining. Functions which configure a Criterion structure will need
+  to be updated accordingly, or will need to be changed to work with the
+  `Benchmark` or `ParameterizedBenchmark` types to do per-benchmark configuration
+  instead.
+- The functions taken by `Criterion.bench_function` and 
+  `Criterion.bench_function_over_inputs` must now have a `'static` lifetime.
+  This means that you may need to change your closures from `|bencher| {...}` 
+  to `move |bencher| {...}`.
+
+### Removed
+- The hidden `criterion::ConfidenceInterval` and`criterion::Estimate` types are
+  no longer publicly accessible.
+- The `Criterion.summarize` function has been removed.
 
 ## [0.1.2]
 ### Changed